split pypsa help functions
