+++ workflow/scripts/prepare_base_network.py
@@ -34,6 +34,8 @@ from functions import HVAC_cost_curve
 from _helpers import (
     configure_logging,
     mock_snakemake,
+)
+from _pypsa_helpers import (
     make_periodic_snapshots,
     shift_profile_to_planning_year,
 )
